first go at cleaning up
This removes the weird image mechanism and uses an AJAX requests
instead. I'm not sure why this didn't occur to me in the first place.

I removed the scores from the sitemap for now. Not sure if they'll come
back. The admin interface should be good enough as a replacement.

Icons are now SVG - no more themeing. You can use CSS for that. Proper
styling in another commit.